+++ src/Cake.Common/Tools/NUnit/NUnitRunner.cs
@@ -124,6 +124,16 @@ namespace Cake.Common.Tools.NUnit
                 builder.Append("/trace:" + settings.Trace);
             }
 
+            if (settings.OutputFile != null)
+            {
+                builder.AppendQuoted(string.Format(CultureInfo.InvariantCulture, "/output:{0}", settings.OutputFile.MakeAbsolute(_environment).FullPath));
+            }
+
+            if (settings.ErrorOutputFile != null)
+            {
+                builder.AppendQuoted(string.Format(CultureInfo.InvariantCulture, "/err:{0}", settings.ErrorOutputFile.MakeAbsolute(_environment).FullPath));
+            }
+
             if (settings.ResultsFile != null && settings.NoResults)
             {
                 throw new ArgumentException(