+++ src/Cake/Scripting/Roslyn/Nightly/RoslynNightlyScriptSession.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.Linq;
 using System.Reflection;
-using Cake.Core;
 using Cake.Core.Diagnostics;
 using Cake.Core.IO;
 using Cake.Core.Scripting;
@@ -28,7 +26,7 @@ namespace Cake.Scripting.Roslyn.Nightly
             _namespaces = new HashSet<string>(StringComparer.Ordinal);
         }
 
-        public void AddReferencePath(FilePath path)
+        public void AddReference(FilePath path)
         {
             if (path == null)
             {
@@ -57,77 +55,20 @@ namespace Cake.Scripting.Roslyn.Nightly
             }
         }
 
-        public void Execute(string code)
+        public void Execute(Script script)
         {
+            // Generate the script code.
+            var generator = new RoslynCodeGenerator();
+            var code = generator.Generate(script);
+
             // Create the script options dynamically.
-            var scriptOptionsDefault = GetDefaultScriptOptionsField();
-            dynamic options = scriptOptionsDefault.GetValue(null);
-            options = options.AddNamespaces(_namespaces);
-            options = options.AddReferences(_references);
-            options = options.AddReferences(_referencePaths.Select(r => r.FullPath));
+            var options = Microsoft.CodeAnalysis.Scripting.ScriptOptions.Default
+                .AddNamespaces(_namespaces)
+                .AddReferences(_references)
+                .AddReferences(_referencePaths.Select(r => r.FullPath));
 
-            // Execute the code.
-            var csharpScriptEvalMethod = GetEvalMethod();
             _log.Debug("Compiling build script...");
-            csharpScriptEvalMethod.Invoke(null, new object[] { code, options, _host });
-        }
-
-        private static Type FindType(string assemblyName, string typeName)
-        {
-            var assemblies = AppDomain.CurrentDomain.GetAssemblies();
-            foreach (var assembly in assemblies)
-            {
-                if (assembly.FullName.StartsWith(string.Concat(assemblyName, ",")))
-                {
-                    var type = GetLoadableTypes(assembly).FirstOrDefault(x => x.FullName == typeName);
-                    if (type != null)
-                    {
-                        return type;
-                    }
-                }
-            }
-            const string format = "Could not find type '{0}' in assembly '{1}'.";
-            throw new CakeException(string.Format(CultureInfo.InvariantCulture, format, typeName, assemblyName));
-        }
-
-        private static IEnumerable<Type> GetLoadableTypes(Assembly assembly)
-        {
-            if (assembly == null)
-            {
-                throw new ArgumentNullException("assembly");
-            }
-            try
-            {
-                return assembly.GetTypes();
-            }
-            catch (ReflectionTypeLoadException e)
-            {
-                return e.Types.Where(t => t != null);
-            }
-        }
-
-        private static MethodInfo GetEvalMethod()
-        {
-            // Get the method.
-            var type = FindType("Microsoft.CodeAnalysis.Scripting.CSharp", "Microsoft.CodeAnalysis.Scripting.CSharp.CSharpScript");
-            var method = type.GetMethods(BindingFlags.Static | BindingFlags.Public)
-                .FirstOrDefault(y => y.Name == "Eval" && y.GetParameters().Length == 3);
-            if (method == null)
-            {
-                throw new CakeException("Could not resolve method 'Microsoft.CodeAnalysis.Scripting.CSharp.CSharpScript.Eval'.");
-            }
-            return method;
-        }
-
-        private static FieldInfo GetDefaultScriptOptionsField()
-        {
-            var scriptOptionsType = FindType("Microsoft.CodeAnalysis.Scripting", "Microsoft.CodeAnalysis.Scripting.ScriptOptions");
-            var scriptOptionsDefault = scriptOptionsType.GetField("Default", BindingFlags.Static | BindingFlags.Public);
-            if (scriptOptionsDefault == null)
-            {
-                throw new CakeException("Could not resolve field 'Microsoft.CodeAnalysis.Scripting.Default'.");
-            }
-            return scriptOptionsDefault;
+            Microsoft.CodeAnalysis.Scripting.CSharp.CSharpScript.Eval(code, options, _host);
         }
     }
 }